Fixing preRequestProcess / url rendering order.
Looking at the way the MySpace provider was manipulating a request
in its preRequestProcess, I saw that it was looking at data in
the request object, and not the URL, which seemed weird since 
MySpace only uses REST and should be manipulating the URL.

I realized that previously the call to constructUrl was after the
preRequestProcess, and preRequestProcess was only expecting the 
service base url to be passed in.

I've changed the behavior here and left comments to make this 
more apparent.